reader: do u32 -> u64 conversion higher up the stack
we're about to implement a file reader which will use this method. since
the u32 bits are a result of fuse, let's move this a bit higher up the
stack so we have to do less casting in that new file reader.

+++ reader/src/puzzlefs.rs
@@ -101,14 +101,14 @@ impl<'a> PuzzleFS<'a> {
         Err(FSError::from_errno(Errno::ENOENT))
     }
 
-    pub fn file_read(&self, inode: &Inode, offset: u64, size: u32) -> FSResult<Vec<u8>> {
+    pub fn file_read(&self, inode: &Inode, offset: u64, size: u64) -> FSResult<Vec<u8>> {
         let chunks = match &inode.mode {
             InodeMode::File { chunks } => chunks,
             _ => return Err(FSError::from_errno(Errno::ENOTDIR)),
         };
 
         // TODO: fix all this casting...
-        let end = offset + u64::from(size);
+        let end = offset + size;
         let mut data = vec![0_u8; size as usize];
 
         let mut file_offset = 0;
